updated code with !! in variables Activity 01 d in Canvas

+++ src/main/kotlin/activity_01_d/Main.kt
@@ -43,7 +43,7 @@ fun main() {
         println("You entered $num5. Okay! Now enter a number to divide all numbers to.")
         println(enterHere)
 
-        var divisor : Double = readLine()!!.toDouble()
+        var divisor : Int = readLine()!!.toInt()
 //        var divisor : Double = readLine().toDouble()
 //        var divisorDouble = divisor.toDouble()
